pyproject.toml
[tool.poetry]
name = "email_notification_microservice"
version = "0.1.0"
description = "A simple email and push notification mircroservie made with django."
authors = ["Naman Tamrakar <namantam1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.8"
django-environ = "^0.7.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.1"
celery = "^5.1.2"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}
ipython = "^7.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# taken from poetry repository
[tool.black]
line-length = 88
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''
